requirements.txt
streamlit==1.17.0
altair==4.2.0
python-dotenv==1.0.1
openai==0.28.0  # Adjust based on your needs
